+++ Examples/Resources/WSManListener/2-WSManListener_HTTPS_Config.ps1
@@ -3,7 +3,7 @@
 .GUID 394755a7-6f2e-4213-ba2b-4688c70b2d0d
 .AUTHOR Daniel Scott-Raynsford
 .COMPANYNAME
-.COPYRIGHT (c) 2018 Daniel Scott-Raynsford. All rights reserved.
+.COPYRIGHT (c) Daniel Scott-Raynsford. All rights reserved.
 .TAGS DSCConfiguration
 .LICENSEURI https://github.com/PlagueHO/WSManDsc/blob/master/LICENSE
 .PROJECTURI https://github.com/PlagueHO/WSManDsc